# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Weather API",
        "description": "Documentación de la API de clima con autenticación, búsquedas y favoritos",
        "contact": {
            "email": "tucorreo@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor API"
        }
    ],
    "paths": {
...
        "/api/searches": {
            "get": {
                "tags": [
                    "Searches"
                ],
                "summary": "Listar todas las búsquedas del usuario autenticado",
                "operationId": "103c52b36fe771ce5e9de13d22104425",
                "responses": {
                    "200": {
                        "description": "Lista de búsquedas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Búsquedas recuperadas exitosamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Searches"
                ],
                "summary": "Crear una nueva búsqueda",
                "operationId": "fbdaf037bdef2a6e9d14020eabc7b925",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "city"
                                ],
                                "properties": {
                                    "city": {
                                        "type": "string",
                                        "example": "Madrid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Búsqueda creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Búsqueda creada"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error al obtener el clima"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/searches/{id}": {
            "get": {
                "tags": [
                    "Searches"
                ],
                "summary": "Mostrar una búsqueda específica",
                "operationId": "b97fa42d613bde00e5da2454a2a19d10",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la búsqueda",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Búsqueda encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Búsqueda encontrada"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Búsqueda no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Searches"
                ],
                "summary": "Actualizar una búsqueda",
                "operationId": "e72a2058f0a94ccfdbace030be365b67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la búsqueda a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "city"
                                ],
                                "properties": {
                                    "city": {
                                        "type": "string",
                                        "example": "Lima"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Búsqueda actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Búsqueda actualizada"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al actualizar"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Searches"
                ],
                "summary": "Eliminar una búsqueda",
                "operationId": "fd13d7f745d5caafaec40ee714ca7400",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la búsqueda a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Búsqueda eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Búsqueda eliminada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al eliminar"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}